requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
matplotlib
